# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Four Delivery API",
        "description": "API para o sistema de delivery Four Delivery",
        "contact": {
            "email": "suporte@fourdelivery.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/admin-v1",
            "description": "Admin API V1"
        },
        {
            "url": "/api/app-v1",
            "description": "App API V1"
        }
    ],
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Listar todas as categorias",
                "operationId": "0d9f6f774f1b480b5318196b0ce0dc72",
                "responses": {
                    "200": {
                        "description": "Lista de categorias",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Criar nova categoria",
                "operationId": "d566fee2a2234738bf97d6ef32889c37",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "company_id",
                                    "name",
                                    "slug",
                                    "description",
                                    "image_url"
                                ],
                                "properties": {
                                    "company_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Combos"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "combos"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Melhores combos da região"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "http://example.com/image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoria criada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categories-with-total-items": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Listar categorias com total de itens e resumo de estoque",
                "operationId": "0ced9750d0a45b2e16c0dddaa817d2ff",
                "responses": {
                    "200": {
                        "description": "Lista de categorias com detalhes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categories-with-company-items": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Listar categorias com itens da empresa",
                "operationId": "ea2d95d02dc6260c53890004796ed896",
                "responses": {
                    "200": {
                        "description": "Lista de categorias e seus itens",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Mostrar detalhes de uma categoria",
                "operationId": "875d4f0851994f12ca2b94f007a0a874",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da categoria e itens paginados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Atualizar uma categoria",
                "operationId": "d2c95300f3f0e703cd4d4b3763e6e010",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "image_url": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoria atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Excluir uma categoria",
                "operationId": "198126e145163c3c4a6d3ada0a7c3f07",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Excluído com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}